striker: Update all the things for Stretch
This role has been broken since the base image update from Jessie to
Stretch. The Stretch package for Keystone runs using Apache
and mod_wsgi rather than as a standalone python service. This creates
a conflict with the way that we were previously deploying Striker
because Keystone needs a python2 runtime.

* Run Keystone as a wsgi process with Apache
* Add a service::uwsgi define
* Run Striker as a uwsgi app
* Update Phabricator configuration for upstream changes
* Convert uwsgi::app systemd unit definition to an epp template
* Patch the upstream Keystone bootstrap process to work with our LDAP
  backend.
* Update wmfkeystonehooks for newer version of Keystone

Change-Id: I37ffcdf95e4ebc65fa9b5105d3be31f06c2b49a1