Move tidy dependencies to the tidy role
Provisioning a new machine is currently failing on a tidy dependency.
According to an FAQ on mw.org,[1], tidy is no longer used in current
versions, so it can be dropped from the mediawiki role, but the
dependencies should be moved to the tidy role for use by backport devs.

[1] https://www.mediawiki.org/wiki/Parsing/Replacing_Tidy/FAQ

Bug: T216490
Change-Id: I3097a51e97bee3afe10591fc58575b678aeeacf7