Avoid error deluge in wikihasextension
The wikihasextension script is used to send a list of DB names to various shell
scripts which typically invoke some MediaWiki command on each of them. If alldbs
prints an error notice or a stack trace on stdout, each word of that output will
be interpreted as a DB name, resulting in a sea of errors, making the original
error very hard to identify.

Make wikihasextension use stderr for error output to avoid that.

Change-Id: Ibd69699c8e856a36748baa4570c9b89d5e8584b5