Remove browser testing concerns from the centralauth role
The centralauth role includes the browsertests class, which is broken
due to T235138. Browser testing concerns seem out of place in this role
in any case, whatever the historical reasons for including them. Also,
Zeljko says that the browser testing setup supported by the browsertests
class is long-since deprecated. So, this patch removes browser testing
concerns from the role.

Bug: T235139
Change-Id: I2a6873cfb9d724aab606c4f5700596e3773acf26