Re-set password after installer finishes
The installer fails to set the password due to policy conflicts.
As a lazy workaround, set it manually.

This is not great as it is a refreshonly resource so it adds one
more thing that needs to be re-done manually if the installer errors
out, but it's better than nothing.

Bug: T232354
Change-Id: I60da31f767cb5b44e75170111580800e00e9c937