Split visualeditor role
Separate the minimum required for setting up VisualEditor and all
the extra features that intend to create an environment similar to
Wikimedia production.

This is a breaking change if you have enabled the visualeditor role
(you'll need to enable the visualeditor_wikimedia role instead).

Also add a missing VisualEditor dependency to the discussiontools
role.

Bug: T260503
Change-Id: I65abb66141de4e13d259d038bb43482bef2889d0