role::oauth: Add wgOAuth2PrivateKey and wgOAuth2PublicKey
Add config settings for the OAuth 2.0 provider:
* wgOAuth2PrivateKey: x509 private key for JWT signing
* wgOAuth2PublicKey: x509 public key for JWT signing

The static values that are provisioned for these settings are "insecure"
in that any MediaWiki-Vagrant managed vm will be using the same values.
The need for this config however seems to be purely to quite log
errors/warnings as signed JWTs are not actually being used in the
MediaWiki integration. Instead they are a requirement of the library
being used to provide OAuth 2.0 support.

Change-Id: I96d0967a365af65159c5f8822ba8b5b16fa7a1e2