EventBus: Disable sending events during PHPUnit tests
A recent change (T262443) caused PHPUnit tests to begin failing if a live
HTTP request was made. This caused failures in tests that incidentally
triggered event submissions.

This change disables EventBus from submitting events in the context of
PHPUnit tests, so as to avoid breaking unrelated tests. This change does
not break EventBus' own tests, as they do not depend on events actually
being sent.

Bug: T270801
Change-Id: I320b789ea243e280a01e2aed34b2ca63527d13d3