Validate IP address for `vagrant config static_ip <ip>`
Previously, `vagrant config` accepted an invalid address, and then
all futher vagrant commands failed because loading the settings
file threw an exception. With this patch, using an invalid address
will still result in an exception, but at least it will happen
when trying to add it to the settings and not afterwards, so it
won't brick the system.

Change-Id: I6938676d8128a1cefe2dfd6521951f9de6946d06