Make number of parallel composer jobs configurable for git-update
The `vagrant git-update` command runs many `composer update` commands
in parallel. Depending on how many extensions are installed and how
the VM is configured, this might lead to OOM errors some or all the
time. Allow opting out of it at the cost of slower updates via
`vagrant hiera mediawiki::parallel_composer_jobs 1` or such.

There is probably a better way to manage parallel jobs than this,
but it beats git-update not working at all.

Bug: T233920
Change-Id: I89771d335f446ed7af24fd571023b48d45b3b40d