change static_ip to conform to VirtualBox 6.1.28 network range restrictions
As of VirtualBox 6.1.28 (released October 19, 2021) host-only network
ranges are restricted to 192.168.56.0/21 (192.168.56.0 - 192.168.63.255).
The default static ip in MediaWiki-Vagrant is set to 10.11.12.13.
VirtualBox quits with error while bringing up a VM with `vagrant up`.
Users can configure the static ip with `vagrant config` as a workaround,
but attempting to override the default ip address by adding a
config.vm.network setting to Vagrantfile-extra.rb fails because Vagrant
treats it as an additional network. To avoid requiring an extra command
just to bring up a new box, we can change the default static ip address
to something within the allowed range.

Bug: T305753
Change-Id: Ie330a2bfe4b6774cad60e1eafce70f0a2a6a503d