Older NPM version should only apply to nodejs6
This workaround is documented to be in place for
nodejs6 specifically. More recent node versions
(e.g. eventgate role requires nodejs10) will
come with more recent NPM versions that don't
appear to suffer from the documented issues, but
forcing it to install an older NPM version in such
case may end up installing one that is not even
compatible with the installed nodejs version.

Change-Id: I85e6830517af889170e4e5b518b079dc56fb2920