Set git user
When user.name and user.email are unset, `vagrant git-update`
fails with "Please tell me who you are." when local patches need
to be rebased.

Do the lazy thing and set it to something fake. Users who prefer
to use git inside the VM can still override it manually.
Gerrit will reject patches with unknown authors so there is no
harm if they forget to do so.

Bug: T86299
Change-Id: Ie8ab6884557dc164da3bc7f0dcfbba19afb4f94e