Implement NVM to service class
Currently the only installed version of node is 12,
this is less than ideal as most production services have migrated to a range of newer versions.
This implementation installs NVM and uses it when setting up service classes,
or for npm::install via a version parameter. Shell calls to node/npm still use version 12
and npm::install defaults to 12 so other parts of mediawiki-vagrant shouldn't break.
Also deleted npm::update as this doesn't seem to have been used.

Change-Id: I28cfa5541a770b0ea139d11b1aa70abfa38af896